basic ui with 3 scenes
basic ui for the two levels as well as a main menu, game over, and win screen. The main menu screen should load the game if you change the level 1 to scene one in the build settings. Let me know if you want me to code anything or change the way the UI looks at all.